--- dump/index.html
@@ -1,103 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Title</title>
-  <style>
-      .preloader {
-          position: fixed;
-          top: 0;
-          left: 0;
-          width: 100%;
-          height: 100vh;
-          display: flex;
-          justify-content: center;
-          align-items: center;
-          background-image: -moz-linear-gradient(89deg, #5f16a9 0%, #cf15f8 100%);
-          background-image: -webkit-linear-gradient(89deg, #5f16a9 0%, #cf15f8 100%);
-          background-image: -ms-linear-gradient(89deg, #5f16a9 0%, #cf15f8 100%);
-          z-index: 999;
-          overflow: hidden; }
-      .preloader .spinner {
-          margin: 100px auto;
-          width: 40px;
-          height: 40px;
-          position: relative;
-          text-align: center;
-          -webkit-animation: sk-rotate 2s infinite linear;
-          animation: sk-rotate 2s infinite linear; }
-      .preloader .dot1, .preloader .dot2 {
-          width: 60%;
-          height: 60%;
-          display: inline-block;
-          position: absolute;
-          top: 0;
-          background-color: #ffffff;
-          border-radius: 100%;
-          -webkit-animation: sk-bounce 2s infinite ease-in-out;
-          animation: sk-bounce 2s infinite ease-in-out; }
-      .preloader .dot2 {
-          top: auto;
-          bottom: 0;
-          -webkit-animation-delay: -1s;
-          animation-delay: -1s; }
-
-      @-webkit-keyframes sk-rotate {
-          100% {
-              -webkit-transform: rotate(360deg);
-              -ms-transform: rotate(360deg);
-              transform: rotate(360deg); } }
-      @-moz-keyframes sk-rotate {
-          100% {
-              -webkit-transform: rotate(360deg);
-              -ms-transform: rotate(360deg);
-              transform: rotate(360deg); } }
-      @-ms-keyframes sk-rotate {
-          100% {
-              -webkit-transform: rotate(360deg);
-              -ms-transform: rotate(360deg);
-              transform: rotate(360deg); } }
-      @keyframes sk-rotate {
-          100% {
-              -webkit-transform: rotate(360deg);
-              -ms-transform: rotate(360deg);
-              transform: rotate(360deg); } }
-      @-webkit-keyframes sk-bounce {
-          0%, 100% {
-              transform: scale(0);
-              -webkit-transform: scale(0); }
-          50% {
-              transform: scale(1);
-              -webkit-transform: scale(1); } }
-      @-moz-keyframes sk-bounce {
-          0%, 100% {
-              transform: scale(0);
-              -webkit-transform: scale(0); }
-          50% {
-              transform: scale(1);
-              -webkit-transform: scale(1); } }
-      @-ms-keyframes sk-bounce {
-          0%, 100% {
-              transform: scale(0);
-              -webkit-transform: scale(0); }
-          50% {
-              transform: scale(1);
-              -webkit-transform: scale(1); } }
-      @keyframes sk-bounce {
-          0%, 100% {
-              transform: scale(0);
-              -webkit-transform: scale(0); }
-          50% {
-              transform: scale(1);
-              -webkit-transform: scale(1); } }
-  </style>
-</head>
-<body>
-<div class="preloader">
-  <div class="spinner">
-    <div class="dot1"></div>
-    <div class="dot2"></div>
-  </div>
-</div>
-</body>
-</html>
