--- dump/pr.scss
@@ -1,63 +1,0 @@
-@mixin gradient-one {
-  background-image: -moz-linear-gradient( 89deg, rgb(95,22,169) 0%, rgb(207,21,248) 100%);
-  background-image: -webkit-linear-gradient( 89deg, rgb(95,22,169) 0%, rgb(207,21,248) 100%);
-  background-image: -ms-linear-gradient( 89deg, rgb(95,22,169) 0%, rgb(207,21,248) 100%);
-}
-
-.preloader {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  @include gradient-one;
-  z-index: 999;
-  overflow: hidden;
-  .spinner {
-    margin: 100px auto;
-    width: 40px;
-    height: 40px;
-    position: relative;
-    text-align: center;
-    -webkit-animation: sk-rotate 2.0s infinite linear;
-    animation: sk-rotate 2.0s infinite linear;
-  }
-  .dot1,
-  .dot2 {
-    width: 60%;
-    height: 60%;
-    display: inline-block;
-    position: absolute;
-    top: 0;
-    background-color: #ffffff;
-    border-radius: 100%;
-
-    -webkit-animation: sk-bounce 2.0s infinite ease-in-out;
-    animation: sk-bounce 2.0s infinite ease-in-out;
-  }
-  .dot2 {
-    top: auto;
-    bottom: 0;
-    -webkit-animation-delay: -1.0s;
-    animation-delay: -1.0s;
-  }
-}
-@include keyframes (sk-rotate) {
-  100% {
-    @include transform(rotate(360deg));
-  }
-}
-@include keyframes (sk-bounce) {
-  0%,
-  100% {
-    transform: scale(0.0);
-    -webkit-transform: scale(0.0);
-  }
-  50% {
-    transform: scale(1.0);
-    -webkit-transform: scale(1.0);
-  }
-}
