+++ src/styles/space.module.scss
