+++ src/utils/func.ts
