+++ src/utils/intefaces.ts
